requirements.txt
numpy==2.2.1
pandas==2.2.3
polars==1.17.1
scipy==1.15.1
biopython==1.84
